requirements.txt
asn1crypto==0.24.0
boto==2.49.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
idna==2.8
psycopg2-binary==2.8.2
pycparser==2.19
python-telegram-bot==12.0.0b1
requests==2.21.0
six==1.12.0
tornado==6.0.2
urllib3==1.24.1
